django_rest_framework/EastcarServer/requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
diff-match-patch==20121119
Django==1.11.6
django-cors-headers==2.1.0
django-crispy-forms==1.7.0
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==0.7.0
django-redis==4.8.0
django-reversion==2.0.13
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
drf-extensions==0.3.1
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.6
itypes==1.1.0
jdcal==1.3
Jinja2==2.10
Markdown==2.6.9
MarkupSafe==1.0
mysqlclient==1.3.12
odfpy==1.3.6
openpyxl==2.5.0
Pillow==5.0.0
pycparser==2.19
pycryptodome==3.4.11
PyJWT==1.5.3
PyMySQL==0.9.2
pytz==2017.2
PyYAML==3.12
redis==2.10.6
requests==2.18.4
six==1.10.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
xlrd==1.1.0
XlsxWriter==1.1.0
xlwt==1.3.0
